+++ src/additionalInfo.ts
@@ -4,7 +4,7 @@ import { clamp, createEnum, layerColors, MOCK_HANDLERS } from './helpers';
 import { LayerAdditionalInfo, BezierPath, Psd, BrightnessAdjustment, ExposureAdjustment, VibranceAdjustment, ColorBalanceAdjustment, BlackAndWhiteAdjustment, PhotoFilterAdjustment, ChannelMixerChannel, ChannelMixerAdjustment, PosterizeAdjustment, ThresholdAdjustment, GradientMapAdjustment, CMYK, SelectiveColorAdjustment, ColorLookupAdjustment, LevelsAdjustmentChannel, LevelsAdjustment, CurvesAdjustment, CurvesAdjustmentChannel, HueSaturationAdjustment, HueSaturationAdjustmentChannel, PresetInfo, Color, ColorBalanceValues, WriteOptions, LinkedFile, PlacedLayerType, Warp, KeyDescriptorItem, BooleanOperation, LayerEffectsInfo, Annotation, LayerVectorMask, AnimationFrame, Timeline, PlacedLayerFilter, UnitsValue, Filter, PlacedLayer, ReadOptions, Layer } from './psd';
 import { PsdReader, readSignature, readUnicodeString, skipBytes, readUint32, readUint8, readFloat64, readUint16, readBytes, readInt16, checkSignature, readFloat32, readFixedPointPath32, readSection, readColor, readInt32, readPascalString, readUnicodeStringWithLength, readAsciiString, readPattern, readLayerInfo } from './psdReader';
 import { PsdWriter, writeZeros, writeSignature, writeBytes, writeUint32, writeUint16, writeFloat64, writeUint8, writeInt16, writeFloat32, writeFixedPointPath32, writeUnicodeString, writeSection, writeUnicodeStringWithPadding, writeColor, writePascalString, writeInt32 } from './psdWriter';
-import { Annt, BlnM, DescriptorColor, DescriptorUnitsValue, parsePercent, parseUnits, parseUnitsOrNumber, QuiltWarpDescriptor, strokeStyleLineAlignment, strokeStyleLineCapType, strokeStyleLineJoinType, TextDescriptor, textGridding, unitsPercent, unitsValue, WarpDescriptor, warpStyle, writeVersionAndDescriptor, readVersionAndDescriptor, StrokeDescriptor, Ornt, horzVrtcToXY, LmfxDescriptor, Lfx2Descriptor, FrameListDescriptor, TimelineDescriptor, FrameDescriptor, xyToHorzVrtc, serializeEffects, parseEffects, parseColor, serializeColor, serializeVectorContent, parseVectorContent, parseTrackList, serializeTrackList, FractionDescriptor, BlrM, BlrQ, SmBQ, SmBM, DspM, UndA, Cnvr, RplS, SphM, Wvtp, ZZTy, Dstr, Chnl, MztT, Lns, blurType, DfsM, ExtT, ExtR, FlCl, CntE, WndM, Drct, IntE, IntC, FlMd, unitsPercentF, frac, ClrS, descBoundsToBounds, boundsToDescBounds, presetKindType, gradientInterpolationMethodType } from './descriptor';
+import { Annt, BlnM, DescriptorColor, DescriptorUnitsValue, parsePercent, parseUnits, parseUnitsOrNumber, QuiltWarpDescriptor, strokeStyleLineAlignment, strokeStyleLineCapType, strokeStyleLineJoinType, TextDescriptor, textGridding, unitsPercent, unitsValue, WarpDescriptor, warpStyle, writeVersionAndDescriptor, readVersionAndDescriptor, StrokeDescriptor, Ornt, horzVrtcToXY, LmfxDescriptor, Lfx2Descriptor, FrameListDescriptor, TimelineDescriptor, FrameDescriptor, xyToHorzVrtc, serializeEffects, parseEffects, parseColor, serializeColor, serializeVectorContent, parseVectorContent, parseTrackList, serializeTrackList, FractionDescriptor, BlrM, BlrQ, SmBQ, SmBM, DspM, UndA, Cnvr, RplS, SphM, Wvtp, ZZTy, Dstr, Chnl, MztT, Lns, blurType, DfsM, ExtT, ExtR, FlCl, CntE, WndM, Drct, IntE, IntC, FlMd, unitsPercentF, frac, ClrS, descBoundsToBounds, boundsToDescBounds, presetKindType, gradientInterpolationMethodType, PxScDescriptor } from './descriptor';
 import { serializeEngineData, parseEngineData } from './engineData';
 import { encodeEngineData, decodeEngineData } from './text';
 import { decodeEngineData2 } from './engineData2';
@@ -975,6 +975,76 @@ addHandler(
 				writeVersionAndDescriptor(writer, '', 'null', desc);
 			}, true);
 		}
+	},
+);
+
+addHandler(
+	'PxSc',
+	() => false,
+	(reader, target) => {
+		const desc = readVersionAndDescriptor(reader, true) as PxScDescriptor;
+		// console.log('PxSc', require('util').inspect(desc, false, 99, true));
+
+		if (desc.pixelSourceType === 1986285651) {
+			target.pixelSource = {
+				type: 'vdPS',
+				origin: { x: desc.origin.Hrzn, y: desc.origin.Vrtc },
+				interpretation: {
+					interpretAlpha: desc.interpretation.interpretAlpha.split('.')[1] as any,
+					profile: desc.interpretation.profile,
+				},
+				frameReader: {
+					type: 'QTFR',
+					link: {
+						name: desc.frameReader['Lnk ']['Nm  '],
+						fullPath: desc.frameReader['Lnk '].fullPath,
+						originalPath: desc.frameReader['Lnk '].originalPath,
+						relativePath: desc.frameReader['Lnk '].relPath,
+						alias: desc.frameReader['Lnk '].alis,
+					},
+					mediaDescriptor: desc.frameReader.mediaDescriptor,
+				},
+				showAlteredVideo: desc.showAlteredVideo,
+			};
+		} else {
+			reader.log(`Unknown pixelSourceType`);
+		}
+	},
+	(writer, target) => {
+		const source = target.pixelSource!;
+		const desc: PxScDescriptor = {
+			_name: '',
+			_classID: 'PixelSource',
+			pixelSourceType: 1986285651, // vdP
+			descVersion: 1,
+			origin: { Hrzn: source.origin.x, Vrtc: source.origin.y },
+			interpretation: {
+				_name: '',
+				_classID: 'footageInterpretation',
+				Vrsn: 1,
+				interpretAlpha: `alphaInterpretation.${source.interpretation.interpretAlpha}` as any,
+				profile: source.interpretation.profile,
+			},
+			frameReader: {
+				_name: '',
+				_classID: 'FrameReader',
+				frameReaderType: 1364477522, // QTF
+				descVersion: 1,
+				'Lnk ': {
+					_name: '',
+					_classID: 'ExternalFileLink',
+					descVersion: 2,
+					'Nm  ': source.frameReader.link.name,
+					fullPath: source.frameReader.link.fullPath,
+					originalPath: source.frameReader.link.originalPath,
+					alis: source.frameReader.link.alias,
+					relPath: source.frameReader.link.relativePath,
+				},
+				mediaDescriptor: source.frameReader.mediaDescriptor,
+			},
+			showAlteredVideo: source.showAlteredVideo,
+		};
+		writeVersionAndDescriptor(writer, '', 'PixelSource', desc);
 	},
 );
 
@@ -3627,6 +3697,8 @@ if (MOCK_HANDLERS) {
 
 /*
 interface CAIDesc {
+	_name: '';
+	_classID: 'null';
 	enab: boolean;
 	generationalGuid: string;
 }
@@ -3636,15 +3708,15 @@ addHandler(
 	() => false,
 	(reader, _target, left) => {
 		const version = readUint32(reader); // 3
-		const desc = readVersionAndDescriptor(reader) as CAIDesc;
+		const desc = readVersionAndDescriptor(reader, true) as CAIDesc;
+		console.log('CAI version', version);
 		console.log('CAI', require('util').inspect(desc, false, 99, true));
-		console.log('CAI', { version });
 		console.log('CAI left', readBytes(reader, left())); // 8 bytes left, all zeroes
 	},
 	(_writer, _target) => {
 	},
 );
-*/
+// */
 
 if (MOCK_HANDLERS) {
 	addHandler(
@@ -3659,11 +3731,25 @@ if (MOCK_HANDLERS) {
 	);
 }
 
+// interface OCIODescriptor {
+// 	_name: '';
+// 	_classID: 'documentColorManagementInfo';
+// 	'Knd ': 'icc';
+// 	ocio_display_view: {
+// 		_name: '';
+// 		_classID: 'viewColorManagementInfo';
+// 		display: string;
+// 		view: string;
+// 	};
+// }
+
 if (MOCK_HANDLERS) {
 	addHandler(
-		'OCIO', // generative tech?
+		'OCIO', // document color management info
 		target => (target as any)._OCIO !== undefined,
 		(reader, target, left) => {
+			// const desc = readVersionAndDescriptor(reader, true) as OCIODescriptor;
+			// console.log('OCIO', require('util').inspect(desc, false, 99, true));
 			(target as any)._OCIO = readBytes(reader, left());
 		},
 		(writer, target) => {
@@ -3672,8 +3758,11 @@ if (MOCK_HANDLERS) {
 	);
 }
 
-// interface GenIDesc {
+// interface GenIDescriptor {
+//  _name: '';
+//  _classID: 'genTechInfo';
 // 	isUsingGenTech: number;
+//  externalModelList?: [];
 // }
 
 if (MOCK_HANDLERS) {
@@ -3681,9 +3770,9 @@ if (MOCK_HANDLERS) {
 		'GenI', // generative tech
 		target => (target as any)._GenI !== undefined,
 		(reader, target, left) => {
+			const desc = readVersionAndDescriptor(reader, true); // as GenIDescriptor;
+			console.log('GenI', require('util').inspect(desc, false, 99, true));
 			(target as any)._GenI = readBytes(reader, left());
-			// const desc = readVersionAndDescriptor(reader) as GenIDesc;
-			// console.log('GenI', require('util').inspect(desc, false, 99, true));
 		},
 		(writer, target) => {
 			writeBytes(writer, (target as any)._GenI);